+++ src/element/index.ts
@@ -1,4 +1,3 @@
 export * from './custom-element';
 export * from './on-event';
-export * from './request-render';
 export * from './set-properties';