+++ src/element/request-render.ts
@@ -1,5 +1,0 @@
-import { CURRENT_ELEMENT } from '../properties/tokens';
-
-export function requestRender(): void {
-  CURRENT_ELEMENT.context.requestRender();
-}