+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './directives';
 export * from './element';
 export * from './properties';
 export * from './template';