+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.3] - Rollup config - 2022/02/07
+### Changed
+- Moved directives build to "estrela/directives".
+
+### Removed
+- Dropped `requestRender` function.
+
 ## [0.3.2] - Element Helpers - 2022/02/06
 ### Added
 - Global onEvent function to create event observables.